refactor(arch): Move research scripts to scripts/internal/
Move run_auction_comparator.py, play_policy_gate.py, and
evaluate_diagnostic_tricks.py to scripts/internal/ to separate
internal research tooling from canonical scripts.

Deprecation wrappers at old paths forward with a warning.
