+++ docs/01_core/ARCHITECTURE.md
@@ -37,14 +37,11 @@ Structure:
 ### `scripts/`
 **Blessed tooling entrypoints.**
 
-Current scripts:
+Canonical scripts:
 - `compare_rollup.py` — Drift detection (compares rollup against baseline fixture)
 - `compare_runs.py` — Run comparison utility with bootstrap statistics
-- `evaluate_diagnostic_tricks.py` — Diagnostic Ridge evaluation
 - `generate_report.py` — Per-run report generator
 - `lint_repo.py` — Repository linter (enforces boundaries and data policy)
-- `play_policy_gate.py` — Play policy stability gate
-- `run_auction_comparator.py` — Auction comparator orchestrator
 - `run_bidless_diagnostics.py` — Bidless feature dataset diagnostics
 - `run_charts.py` — Production chart generation (extracted from `reporting.chart_runner`)
 - `run_notebooks.py` — Notebook execution via papermill (CI tooling)
@@ -55,6 +52,15 @@ Current scripts:
 - `train_olsa.py` — OLSa model training (extracted from `models.train_olsa`)
 - `validate_configs.py` — Config validation utility
 - `validate_teacher_roster.py` — Teacher roster validation
+
+### `scripts/internal/`
+**Research and internal tooling.** Not part of the canonical workflow.
+
+- `evaluate_diagnostic_tricks.py` — Diagnostic Ridge evaluation
+- `play_policy_gate.py` — Play policy stability gate
+- `run_auction_comparator.py` — Auction comparator orchestrator
+
+Deprecation wrappers at old paths (`scripts/*.py`) forward to `scripts/internal/` with a warning.
 
 ### `tests/`
 **Test suite only.**
@@ -158,14 +164,18 @@ Every experiment run creates a standardized directory under `data/runs/<run_id>/
 | `scripts/validate_configs.py` | Config validation |
 | `scripts/validate_teacher_roster.py` | Teacher roster validation |
 
-### Research/internal tooling
+### Internal tooling (`scripts/internal/`)
 
 | Command | Purpose |
 |---------|---------|
-| `scripts/run_auction_comparator.py` | Auction comparator orchestrator |
-| `scripts/run_charts.py` | Production chart generation |
-| `scripts/evaluate_diagnostic_tricks.py` | Diagnostic Ridge evaluation |
-| `scripts/play_policy_gate.py` | Play policy stability gate |
+| `scripts/internal/run_auction_comparator.py` | Auction comparator orchestrator |
+| `scripts/internal/evaluate_diagnostic_tricks.py` | Diagnostic Ridge evaluation |
+| `scripts/internal/play_policy_gate.py` | Play policy stability gate |
+
+### Research tooling (canonical path)
+
+| Command | Purpose |
+|---------|---------|
 | `scripts/run_bidless_diagnostics.py` | Bidless diagnostics |
 | `scripts/train_b0.py` | B0 hand value model training |
 | `scripts/train_bidder.py` | Bidder model training |
